LUN-4: Move files to src folder
