+++ __tests__/components/App-test.tsx
@@ -4,7 +4,7 @@
 
 import 'react-native';
 import React from 'react';
-import App from '../../App';
+import App from '../../src/App';
 import {shallow} from 'enzyme';
 
 describe('App', () => {
