--- react-native.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  project: {
-    ios: {},
-    android: {},
-  },
-  assets: ['./assets/fonts/'],
-};
