+++ src/components/Modal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text, TouchableOpacity, Modal} from 'react-native';
 import {IConfirmationModalProps} from '../interfaces/exercise';
-import {CloseIcon} from '../assets/images';
+import {CloseIcon} from '../../assets/images';
 import Button from './Button';
 import {BUTTONS_THEME} from '../constants/data';
 import AsyncStorage from '@react-native-async-storage/async-storage';
