+++ src/components/PopoverContent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {View, Text} from 'react-native';
-import {InfoIcon} from '../assets/images';
+import {InfoIcon} from '../../assets/images';
 import {StyleSheet} from 'react-native';
 import {widthPercentageToDP as wp} from 'react-native-responsive-screen';
 import {COLORS} from '../constants/colors';
