+++ src/interfaces/exercise.ts
