+++ src/interfaces/exercises.ts
