+++ src/screens/Exercises/ExercisesScreens.tsx
@@ -8,17 +8,17 @@ import {
   Pressable,
 } from 'react-native';
 import {IExercisesScreenProps} from '../../interfaces/exercises';
-import {Home} from '../../assets/images';
+import {Home} from '../../../assets/images';
 import Title from '../../components/Title';
 import {EXERCISES, SCREENS} from '../../constants/data';
-import {Arrow} from '../../assets/images';
+import {Arrow} from '../../../assets/images';
 import {useFocusEffect} from '@react-navigation/native';
 import {COLORS} from '../../constants/colors';
 import {
   BackButton,
   BackArrowPressed,
   HomeButtonPressed,
-} from '../../assets/images';
+} from '../../../assets/images';
 import {StyleSheet} from 'react-native';
 import {widthPercentageToDP as wp} from 'react-native-responsive-screen';
