+++ src/screens/Exercises/VocabularyTrainerExerciseScreen.tsx
@@ -17,7 +17,7 @@ import {ENDPOINTS} from '../../constants/endpoints';
 import {IVocabularyTrainerScreen} from '../../interfaces/exercise';
 import AnswerSection from '../../components/AnswerSection';
 import {KeyboardAwareScrollView} from 'react-native-keyboard-aware-scroll-view';
-import {CloseButton} from '../../assets/images';
+import {CloseButton} from '../../../assets/images';
 import {SCREENS} from '../../constants/data';
 import {useFocusEffect} from '@react-navigation/native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
