+++ src/screens/SummaryScreens/InitialSummaryScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text, StatusBar} from 'react-native';
 import Button from '../../components/Button';
-import {CheckIcon, ListIcon, RepeatIcon} from '../../assets/images';
+import {CheckIcon, ListIcon, RepeatIcon} from '../../../assets/images';
 import {BUTTONS_THEME, SCREENS} from '../../constants/data';
 import {IInitialSummaryScreenProps} from '../../interfaces/summaryScreens';
 import {useFocusEffect} from '@react-navigation/native';
