+++ src/screens/SummaryScreens/ResultScreen.tsx
@@ -4,7 +4,11 @@ import {IResultScreenProps} from '../../interfaces/summaryScreens';
 import {useFocusEffect} from '@react-navigation/native';
 import {SCREENS, BUTTONS_THEME, RESULT_PRESETS} from '../../constants/data';
 import {COLORS} from '../../constants/colors';
-import {CircularFinishIcon, NextArrow, RepeatIcon} from '../../assets/images';
+import {
+  CircularFinishIcon,
+  NextArrow,
+  RepeatIcon,
+} from '../../../assets/images';
 import Title from '../../components/Title';
 import Loading from '../../components/Loading';
 import VocabularyOverviewListItem from '../../components/VocabularyOverviewListItem';
